% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACFunctions.R
\name{get}
\alias{get}
\title{Get a parameter}
\usage{
get(object, name)
}
\arguments{
\item{object}{Input simATACCount object.}

\item{name}{Name of the parameter.}
}
\value{
Value of the input parameter.
}
\description{
Get the value of a single parameter from input simATACCount object.
}
