% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{getBin}
\alias{getBin}
\title{Get the name of the input region from bed file in the format of chr:start-end as a string. Start
and end are coordinates of the beginning and end of a specific bin with 5000 length base pairs.}
\usage{
getBin(region, bin.name)
}
\arguments{
\item{region}{Input region in the format of [chr, start, end].}

\item{bin.name}{List of bin names of the raw bin by cell matrix.}
}
\value{
List of bin names that has intersection with the input region.
}
\description{
Get the name of the input region from bed file in the format of chr:start-end as a string. Start
and end are coordinates of the beginning and end of a specific bin with 5000 length base pairs.
}
