% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACSimulate.R
\name{getBinNames}
\alias{getBinNames}
\title{Set bin names based on the species parameter of the simATACCount object.}
\usage{
getBinNames(object)
}
\arguments{
\item{object}{simATACCount object with simulation parameters.}
}
\value{
List of bin names.
}
\description{
Set bin names based on the species parameter of the simATACCount object.
}
