% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{getCountFromh5}
\alias{getCountFromh5}
\title{Extract count matrix from h5 file}
\usage{
getCountFromh5(file)
}
\arguments{
\item{file}{Input .h5 (or snap) file to read cell by bin matrix from}
}
\value{
The cell by bin count matrix from .h5 file
}
\description{
Extract count matrix from h5 file
}
