% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{getCountFromSCE}
\alias{getCountFromSCE}
\title{Return a sparse count matrix from a SingleCellExperiment object. If count matrix is missing
a warning is printed and the first assay is returned.}
\usage{
getCountFromSCE(sce)
}
\arguments{
\item{sce}{SingleCellExperiment input object containing counts.}
}
\value{
A sparse matrix containing counts.
}
\description{
Return a sparse count matrix from a SingleCellExperiment object. If count matrix is missing
a warning is printed and the first assay is returned.
}
