% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACFunctions.R
\name{getParameters}
\alias{getParameters}
\title{Get parameters}
\usage{
getParameters(object, names)
}
\arguments{
\item{object}{Input simATACCount object.}

\item{names}{List of parameter names.}
}
\value{
List of parameter values.
}
\description{
Get the value of input parameters from simATACCount object.
}
