% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACFunctions.R
\name{newsimATACCount}
\alias{newsimATACCount}
\title{newsimATACCount}
\usage{
newsimATACCount(...)
}
\arguments{
\item{...}{Variables to set simATACCount object parameters.}
}
\value{
The new object from class simATACCount.
}
\description{
Create a new simATACCount object.
}
