% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACSimulate.R
\name{setDefautParameters}
\alias{setDefautParameters}
\title{Calculate the binomial coefficients from an existing dataset GSE99172,
for default values.}
\usage{
setDefautParameters(object)
}
\arguments{
\item{object}{simATACCount object with simulation parameters.}
}
\value{
simATACCount object variable with default variables.
}
\description{
Calculate the binomial coefficients from an existing dataset GSE99172,
for default values.
}
