% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACFunctions.R
\name{setParameters}
\alias{setParameters}
\title{Set parameters}
\usage{
setParameters(object, update = NULL, ...)
}
\arguments{
\item{object}{Input simATACCount object.}

\item{update}{Parameter name and value to update in the object variable.}

\item{...}{Parameters to set to object variable.}
}
\value{
simATACCount object with updated parameters.
}
\description{
Set input parameters of the input simATACCount object.
}
