% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{simATACCompare}
\alias{simATACCompare}
\title{Compare the main parameters of simATAC's simulated bin by cell matrix to real
matrix, including library size, bin mean, non-zero cell proportion, and the
relation between bin means and non-zero cell proportions.}
\usage{
simATACCompare(sim, real, address, name)
}
\arguments{
\item{sim}{Simulated bin by cell sparse matrix.}

\item{real}{Real bin by cell sparse matrix.}

\item{address}{Folder address to save plots.}

\item{name}{Name of the data.}
}
\value{

}
\description{
Compare the main parameters of simATAC's simulated bin by cell matrix to real
matrix, including library size, bin mean, non-zero cell proportion, and the
relation between bin means and non-zero cell proportions.
}
