% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACCountClass.R
\docType{class}
\name{simATACCount}
\alias{simATACCount}
\alias{simATACCount-class}
\title{The simATACCount class}
\description{
S4 class that holds parameters for the count matrix of simATAC simulation.
}
\section{Parameters}{


simATAC simulation parameters:

\describe{
    \item{\code{nBins}}{The bin number to simulate.}
    \item{\code{nCells}}{The cell number to simulate.}
    \item{\code{[seed]}}{Seed to use for generating random numbers.}
    \item{\code{[default]}}{The logical variable whether to use default parameters
     (TRUE) or learn from data (FALSE)}
     \item{\code{[species]}}{An string indicating the species of the input cells. 
     Supports "hg38", "hg19", "mm9", and "mm10" in the current version.}
    \item{\emph{Library size parameters}}{
        \describe{
            \item{\code{lib.mean1}}{Mean parameter for the first component of library 
            size bimodal Gaussian distribution.}
            \item{\code{lib.mean2}}{Mean parameter for the second component of library 
            size bimodal Gaussian distribution}
            \item{\code{lib.sd1}}{Standard deviation parameter for the first component 
            of library size bimodal Gaussian distribution.}
            \item{\code{lib.sd2}}{Standard deviation parameter for the second component 
            of library size bimodal Gaussian distribution.}
            \item{\code{lib.prob}}{Probability parameter for the first component in bimodal 
            Gaussian distribution. The probability for the second component is 1-lib.prob.}
        }
    }
    \item{\emph{Non-zero cell proportion parameters}}{
        \describe{
            \item{\code{non.zero.pro}}{The proportion of non-zero cells per bin in the 
            original count matrix}
            \item{\code{mean.coef0}}{Estimated coefficient of power zero variable in the 
            polynomial function}
            \item{\code{mean.coef1}}{Estimated coefficient of power one variable in the 
            polynomial function}
            \item{\code{mean.coef2}}{Estimated coefficient of power two variable in the 
            polynomial function}
        }
    }
    \item{\emph{[noise]}}{
        \describe{
            \item{\code{[noise.mean]}}{Gaussian mean to be added as noise to the final 
            simulated counts}
            \item{\code{[noise.sd]}}{Gaussian standard deviation to be added as noise to 
            the final simulated counts}
        }
    }
    \item{\code{sparse.fac}}{Sparsit factor to be multiplied to the input of Poisson 
      distribution on the final simulated count matrix}
}

The parameters not shown in brackets can be estimated from real data using
\code{\link{simATACEstimate}}. For details of the simATAC simulation see \code{\link{simATACSimulate}}.
}

