% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACEstimate.R
\name{simATACEstBinMean}
\alias{simATACEstBinMean}
\title{Estimate bin coverage mean}
\usage{
simATACEstBinMean(count, object)
}
\arguments{
\item{count}{A sparse count matrix to estimate parameters from.}

\item{object}{simATACCount object to store estimated parameters.}
}
\value{
simATACCount object with updated polynomial function parameters.
}
\description{
Estimate the coefficients of second degree polynomial relation between bins' non-zero
cell proportions and means.
}
\details{
Parameters for the second degree polynomial function between bin means and bin
non-zero cell proportion are estimated by using \code{\link[stats]{lm}} and
\code{\link[stats]{poly}}. simATAC excludes bins having non-zero cell proportion
more than 0.8 in modelling the polynomial function.
}
