% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACEstimate.R
\name{simATACEstimate}
\alias{simATACEstimate}
\alias{simATACEstimate.SingleCellExperiment}
\alias{simATACEstimate.dgCMatrix}
\title{Estimate simATAC simulation parameters}
\usage{
simATACEstimate(count, object = newsimATACCount(), verbose = TRUE)

\method{simATACEstimate}{SingleCellExperiment}(count, object = newsimATACCount(), verbose = TRUE)

\method{simATACEstimate}{dgCMatrix}(count, object = newsimATACCount(), verbose = TRUE)
}
\arguments{
\item{count}{Either a sparse bin by cell count matrix, or a SingleCellExperiment
object containing count matrix to estimate parameters from.}

\item{object}{simATACCount object to store estimated parameters and
count matrix in it.}

\item{verbose}{logical variable. Prints the simulation progress if TRUE.}
}
\value{
simATACCount object containing the estimated counts and parameters.
}
\description{
Estimate parameters for the simATAC simulation from a real bin by cell
input matrix.
}
\seealso{
\code{\link{simATACEstLibSize}}, \code{\link{simATACEstNonZeroPro}}
\code{\link{simATACEstBinMean}}
}
