% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACEstimate.R
\name{simATACEstLibSize}
\alias{simATACEstLibSize}
\title{Estimate library size parameters}
\usage{
simATACEstLibSize(count, object)
}
\arguments{
\item{count}{A sparse count matrix to estimate parameters from.}

\item{object}{simATACCount object to store estimated parameters.}
}
\value{
simATACCount object with updated library size parameter.
}
\description{
A Gaussian mixture distribution is fitted to the log2 transformation
of library size values. The estimated parameters are added to the object
variable. See \code{\link[mixtools]{normalmixEM}} for details on the fitting.
}
