% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACEstimate.R
\name{simATACEstNonZeroPro}
\alias{simATACEstNonZeroPro}
\title{Estimate bins' non-zero cell proportion}
\usage{
simATACEstNonZeroPro(count, object)
}
\arguments{
\item{count}{A sparse count matrix to estimate parameters from.}

\item{object}{simATACCount object to store estimated parameters.}
}
\value{
simATACCount object with updated non-zero cell proportion parameter.
}
\description{
Extract the non-zero cell proportion of each bin among all single cells from
the input count matrix.
}
\details{
Vector of non-zero cell proportions of bins is calculated by
dividing the number of non-zero entries over the number of all cells
for each bin.
}
