% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACFunctions.R
\name{simATACget}
\alias{simATACget}
\title{Get a simATACCount parameter}
\usage{
simATACget(object, name)
}
\arguments{
\item{object}{Input simATACCount object.}

\item{name}{Name of the parameter.}
}
\value{
Value of the input parameter.
}
\description{
Get the value of a single variable from input simATACCount object.
}
