% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{simATACgetBinary}
\alias{simATACgetBinary}
\title{Convert raw bin by cell matrix in an SingleCellExperiment object into the binary version.}
\usage{
simATACgetBinary(sim)
}
\arguments{
\item{sim}{SingleCellExperiment object containing simulation parameters.}
}
\value{
Sparse matrix containing binary version of the input simulated bin by cell matrix.
}
\description{
Convert raw bin by cell matrix in an SingleCellExperiment object into the binary version.
}
