% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{simATACgetCellByPeak}
\alias{simATACgetCellByPeak}
\title{Convert raw bin by cell matrix in an SingleCellExperiment object into the
peak by cell matrix by extracting bins having the highest means.}
\usage{
simATACgetCellByPeak(sim, peak.num = 5000)
}
\arguments{
\item{sim}{SingleCellExperiment object containing simulation parameters.}

\item{peak.num}{Number of peak bins to extract from original bin by cell matrix.}
}
\value{
Sparse matrix containing peak.num bins with the highest bin means.
}
\description{
Convert raw bin by cell matrix in an SingleCellExperiment object into the
peak by cell matrix by extracting bins having the highest means.
}
