% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{simATACgetCellByRegion}
\alias{simATACgetCellByRegion}
\title{Convert raw bin by cell matrix in an SingleCellExperiment object into the region by cell
matrix, with regions defined in the input bed file.}
\usage{
simATACgetCellByRegion(sim, file.bed)
}
\arguments{
\item{sim}{SingleCellExperiment object containing simulation parameters.}

\item{file.bed}{Bed file containing chromosome and start and end positions of regions.}
}
\value{
Sparse matrix containing bins having intersection with the regions in the bed file.
}
\description{
Convert raw bin by cell matrix in an SingleCellExperiment object into the region by cell
matrix, with regions defined in the input bed file.
}
