% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACSimulate.R
\name{simATACSimBinMean}
\alias{simATACSimBinMean}
\title{Simulate bin means from estimated polynomial coefficients using
\code{\link{simBinMeans}}.}
\usage{
simATACSimBinMean(object, sim)
}
\arguments{
\item{object}{simATACCount object with simulation parameters.}

\item{sim}{SingleCellExperiment object containing simulation parameters.}
}
\value{
SingleCellExperiment object with updated simulated bin means.
}
\description{
Simulate bin means from estimated polynomial coefficients using
\code{\link{simBinMeans}}.
}
