% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACSimulate.R
\name{simATACSimLibSize}
\alias{simATACSimLibSize}
\title{Simulate library sizes by sampling from a bimodal Gaussian mixture model based
on the estimated mus and sigmas and weight of its components.}
\usage{
simATACSimLibSize(object, sim)
}
\arguments{
\item{object}{simATACCount object with simulation parameters.}

\item{sim}{SingleCellExperiment object containing simulation parameters.}
}
\value{
SingleCellExperiment object with updated simulation library size.
}
\description{
Simulate library sizes by sampling from a bimodal Gaussian mixture model based
on the estimated mus and sigmas and weight of its components.
}
