% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACSimulate.R
\name{simATACSimTrueCount}
\alias{simATACSimTrueCount}
\title{Adjust the simulated library size and bin means to generate final counts
using Poisson distribution, with optional Gaussian noise.}
\usage{
simATACSimTrueCount(object, sim)
}
\arguments{
\item{object}{simATACCount object with simulation parameters.}

\item{sim}{SingleCellExperiment object containing simulation parameters.}
}
\value{
SingleCellExperiment with simulated final counts.
}
\description{
Adjust the simulated library size and bin means to generate final counts
using Poisson distribution, with optional Gaussian noise.
}
