% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACSimulate.R
\name{simATACSimulate}
\alias{simATACSimulate}
\title{simATAC simulation}
\usage{
simATACSimulate(object = newsimATACCount(), verbose = TRUE, ...)
}
\arguments{
\item{object}{simATACCount object with simulation parameters.
See \code{\link{simATACCount}} for details.}

\item{verbose}{logical variable. Prints the simulation progress if TRUE.}

\item{...}{any additional parameter settings to override what is provided in
\code{simATACCount} object.}
}
\value{
SingleCellExperiment object containing the simulated counts.
}
\description{
Simulate bin by cell count matrix from a sparse single-cell ATAC-seq bin by cell
input using simATAC methods.
}
\details{
simATAC provides the option to manually adjust each of the \code{simATACCount}
object parameter by calling \code{\link{setParameters}}. See examples for a demonstration
of how this can be used.

The simulation involves the following steps:
\enumerate{
    \item Set up simulation object
    \item Simulate library sizes
    \item Simulate non zero entries
    \item Simulate bin means
    \item Create final synthetic counts
}

The final output is a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object that
contains the simulated count matrix. The parameters are stored in the \code{\link{colData}}
(for cell specific information), \code{\link{rowData}} (for bin specific information) or
\code{\link{assays}} (for bin by cell matrix) slots. This additional information includes:
\describe{
    \item{\code{rowData}}{
        \describe{
            \item{BinMean}{The simulated bin means.}
        }
    }
    \item{\code{colData}}{
        \describe{
            \item{LibSize}{The simulated library size values.}
        }
    }
    \item{\code{assays}}{
        \describe{
            \item{counts}{The sparse matrix containing simulated counts.}
        }
    }
}
}
\references{
Navidi Ghaziani Z. (2020).

Paper: \url{}

Code: \url{}
}
\seealso{
\code{\link{simATACSimLibSize}}, \code{\link{simATACSimZeroEntry}},
\code{\link{simATACSimBinMean}}, \code{\link{simATACSimTrueCount}}
}
