% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simATACSimulate.R
\name{simATACSimZeroEntry}
\alias{simATACSimZeroEntry}
\title{Simulate zero and non-zero entries of simATAC simulation count matrix.
The extracted non-zero proportion from original bin by cell matrix is
used as probability to set cells to zero using Bernoulli distribution.}
\usage{
simATACSimZeroEntry(object)
}
\arguments{
\item{object}{simATACCount object with simulation parameters.}
}
\value{
simATACCount object with updated simulation parameters.
}
\description{
Simulate zero and non-zero entries of simATAC simulation count matrix.
The extracted non-zero proportion from original bin by cell matrix is
used as probability to set cells to zero using Bernoulli distribution.
}
