% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{simBinMeans}
\alias{simBinMeans}
\title{Simulate a bin's mean from its non-zero cell proportion
from estimated second degree polynomial coefficients.}
\usage{
simBinMeans(x, c0, c1, c2)
}
\arguments{
\item{x}{non-zero cell proportion of a bin.}

\item{c0}{coefficient of x power 0.}

\item{c1}{coefficient of x power 1.}

\item{c2}{coefficient of x power 2.}
}
\value{
dependent variable y, which is the simulated bin mean.
}
\description{
Simulate a bin's mean from its non-zero cell proportion
from estimated second degree polynomial coefficients.
}
