% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFunctions.R
\name{simATACCompare}
\alias{simATACCompare}
\title{Compare the main parameters of simATAC's simulated bin by cell matrix to real
matrix, including library size, bin mean, non-zero cell proportion, the
relation between bin means and non-zero cell proportions in real and simulated
matrices, and the Pearson correlation of the bin mean and non-zero cell proporton 
parameters in the real and simulated count matrices (if the number of simulated 
cells equals the number of real cells).}
\usage{
simATACCompare(sim, real, address, name)
}
\arguments{
\item{sim}{Simulated bin by cell sparse matrix.}

\item{real}{Real bin by cell sparse matrix.}

\item{address}{Folder address to save plots.}

\item{name}{Name of the data.}
}
\description{
Compare the main parameters of simATAC's simulated bin by cell matrix to real
matrix, including library size, bin mean, non-zero cell proportion, the
relation between bin means and non-zero cell proportions in real and simulated
matrices, and the Pearson correlation of the bin mean and non-zero cell proporton 
parameters in the real and simulated count matrices (if the number of simulated 
cells equals the number of real cells).
}
